Add attendance schedule for MastersPreparationApplication. Change some previous logic

+++ src/main/java/com/example/poc/masterspreparation/service/CustomerServiceImpl.java
@@ -1,12 +1,17 @@
 package com.example.poc.masterspreparation.service;
 
+import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.example.poc.masterspreparation.dto.CustomerDto;
+import com.example.poc.masterspreparation.exception.CustomerCreationException;
 import com.example.poc.masterspreparation.model.Customer;
 import com.example.poc.masterspreparation.model.Role;
 import com.example.poc.masterspreparation.repository.CustomerRepository;
+import com.example.poc.masterspreparation.repository.RoleRepository;
 import com.example.poc.masterspreparation.service.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
@@ -15,6 +20,7 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -25,6 +31,7 @@ import javax.transaction.Transactional;
 public class CustomerServiceImpl implements CustomerService, UserDetailsService {
 
     private final CustomerRepository clientRepository;
+    private final RoleRepository roleRepository;
 
     @Override
     public List<Customer> getAllCustomers() {
@@ -32,8 +39,29 @@ public class CustomerServiceImpl implements CustomerService, UserDetailsService 
     }
 
     @Override
-    public Customer saveCustomer(Customer customer) {
-        return clientRepository.save(customer);
+    public Customer saveCustomer(CustomerDto customer) {
+        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+
+        if (customer.getFirstName().equals("") || customer.getLastName().equals("")
+                || customer.getEmail().equals("") || customer.getPassword().equals("")) {
+            throw new CustomerCreationException("Всі поля клієнта повинні бути заповнені");
+        }
+
+        Customer customerModel = new Customer();
+        customerModel.setFirstName(customer.getFirstName());
+        customerModel.setLastName(customer.getLastName());
+        customerModel.setPassword(bCryptPasswordEncoder.encode(customer.getPassword()));
+        customerModel.setEmail(customer.getEmail());
+
+
+        if (customer.getIsAdmin()) {
+            customerModel.setRoles(Collections.singletonList(roleRepository.findByName("ROLE_ADMIN")));
+            customerModel.setAdmin(true);
+        }
+         else {
+            customerModel.setRoles(Collections.singletonList(roleRepository.findByName("ROLE_USER")));
+        }
+        return clientRepository.save(customerModel);
     }
 
     @Override
@@ -48,7 +76,10 @@ public class CustomerServiceImpl implements CustomerService, UserDetailsService 
 
     @Override
     public void deleteCustomerById(Long id) {
-        clientRepository.deleteById(id);
+        Customer customer = getCustomerById(id);
+        if (!customer.isAdmin()) {
+            clientRepository.deleteById(id);
+        }
     }
 
     public Customer findByEmail(String email) {
